refactore: Move test to seprate module to can be run separetly
