+++ src/quicklist.c
@@ -1789,1546 +1789,3 @@ void quicklistBookmarksClear(quicklist *ql) {
     /* NOTE: We do not shrink (realloc) the quick list. main use case for this
      * function is just before releasing the allocation. */
 }
-
-/* The rest of this file is test cases and test helpers. */
-#ifdef SERVER_TEST
-#include <stdint.h>
-#include <sys/time.h>
-#include "testhelp.h"
-#include <stdlib.h>
-
-#define yell(str, ...) printf("ERROR! " str "\n\n", __VA_ARGS__)
-
-#define ERROR                                                                  \
-    do {                                                                       \
-        printf("\tERROR!\n");                                                  \
-        err++;                                                                 \
-    } while (0)
-
-#define ERR(x, ...)                                                            \
-    do {                                                                       \
-        printf("%s:%s:%d:\t", __FILE__, __func__, __LINE__);                   \
-        printf("ERROR! " x "\n", __VA_ARGS__);                                 \
-        err++;                                                                 \
-    } while (0)
-
-#define TEST(name) printf("test — %s\n", name);
-#define TEST_DESC(name, ...) printf("test — " name "\n", __VA_ARGS__);
-
-#define QL_TEST_VERBOSE 0
-
-#define UNUSED(x) (void)(x)
-static void ql_info(quicklist *ql) {
-#if QL_TEST_VERBOSE
-    printf("Container length: %lu\n", ql->len);
-    printf("Container size: %lu\n", ql->count);
-    if (ql->head)
-        printf("\t(zsize head: %lu)\n", lpLength(ql->head->entry));
-    if (ql->tail)
-        printf("\t(zsize tail: %lu)\n", lpLength(ql->tail->entry));
-    printf("\n");
-#else
-    UNUSED(ql);
-#endif
-}
-
-/* Return the UNIX time in microseconds */
-static long long ustime(void) {
-    struct timeval tv;
-    long long ust;
-
-    gettimeofday(&tv, NULL);
-    ust = ((long long)tv.tv_sec) * 1000000;
-    ust += tv.tv_usec;
-    return ust;
-}
-
-/* Return the UNIX time in milliseconds */
-static long long mstime(void) { return ustime() / 1000; }
-
-/* Iterate over an entire quicklist.
- * Print the list if 'print' == 1.
- *
- * Returns physical count of elements found by iterating over the list. */
-static int _itrprintr(quicklist *ql, int print, int forward) {
-    quicklistIter *iter =
-        quicklistGetIterator(ql, forward ? AL_START_HEAD : AL_START_TAIL);
-    quicklistEntry entry;
-    int i = 0;
-    int p = 0;
-    quicklistNode *prev = NULL;
-    while (quicklistNext(iter, &entry)) {
-        if (entry.node != prev) {
-            /* Count the number of list nodes too */
-            p++;
-            prev = entry.node;
-        }
-        if (print) {
-            int size = (entry.sz > (1<<20)) ? 1<<20 : entry.sz;
-            printf("[%3d (%2d)]: [%.*s] (%lld)\n", i, p, size,
-                   (char *)entry.value, entry.longval);
-        }
-        i++;
-    }
-    quicklistReleaseIterator(iter);
-    return i;
-}
-static int itrprintr(quicklist *ql, int print) {
-    return _itrprintr(ql, print, 1);
-}
-
-static int itrprintr_rev(quicklist *ql, int print) {
-    return _itrprintr(ql, print, 0);
-}
-
-#define ql_verify(a, b, c, d, e)                                               \
-    do {                                                                       \
-        err += _ql_verify((a), (b), (c), (d), (e));                            \
-    } while (0)
-
-static int _ql_verify_compress(quicklist *ql) {
-    int errors = 0;
-    if (quicklistAllowsCompression(ql)) {
-        quicklistNode *node = ql->head;
-        unsigned int low_raw = ql->compress;
-        unsigned int high_raw = ql->len - ql->compress;
-
-        for (unsigned int at = 0; at < ql->len; at++, node = node->next) {
-            if (node && (at < low_raw || at >= high_raw)) {
-                if (node->encoding != QUICKLIST_NODE_ENCODING_RAW) {
-                    yell("Incorrect compression: node %d is "
-                         "compressed at depth %d ((%u, %u); total "
-                         "nodes: %lu; size: %zu; recompress: %d)",
-                         at, ql->compress, low_raw, high_raw, ql->len, node->sz,
-                         node->recompress);
-                    errors++;
-                }
-            } else {
-                if (node->encoding != QUICKLIST_NODE_ENCODING_LZF &&
-                    !node->attempted_compress) {
-                    yell("Incorrect non-compression: node %d is NOT "
-                         "compressed at depth %d ((%u, %u); total "
-                         "nodes: %lu; size: %zu; recompress: %d; attempted: %d)",
-                         at, ql->compress, low_raw, high_raw, ql->len, node->sz,
-                         node->recompress, node->attempted_compress);
-                    errors++;
-                }
-            }
-        }
-    }
-    return errors;
-}
-
-/* Verify list metadata matches physical list contents. */
-static int _ql_verify(quicklist *ql, uint32_t len, uint32_t count,
-                      uint32_t head_count, uint32_t tail_count) {
-    int errors = 0;
-
-    ql_info(ql);
-    if (len != ql->len) {
-        yell("quicklist length wrong: expected %d, got %lu", len, ql->len);
-        errors++;
-    }
-
-    if (count != ql->count) {
-        yell("quicklist count wrong: expected %d, got %lu", count, ql->count);
-        errors++;
-    }
-
-    int loopr = itrprintr(ql, 0);
-    if (loopr != (int)ql->count) {
-        yell("quicklist cached count not match actual count: expected %lu, got "
-             "%d",
-             ql->count, loopr);
-        errors++;
-    }
-
-    int rloopr = itrprintr_rev(ql, 0);
-    if (loopr != rloopr) {
-        yell("quicklist has different forward count than reverse count!  "
-             "Forward count is %d, reverse count is %d.",
-             loopr, rloopr);
-        errors++;
-    }
-
-    if (ql->len == 0 && !errors) {
-        return errors;
-    }
-
-    if (ql->head && head_count != ql->head->count &&
-        head_count != lpLength(ql->head->entry)) {
-        yell("quicklist head count wrong: expected %d, "
-             "got cached %d vs. actual %lu",
-             head_count, ql->head->count, lpLength(ql->head->entry));
-        errors++;
-    }
-
-    if (ql->tail && tail_count != ql->tail->count &&
-        tail_count != lpLength(ql->tail->entry)) {
-        yell("quicklist tail count wrong: expected %d, "
-             "got cached %u vs. actual %lu",
-             tail_count, ql->tail->count, lpLength(ql->tail->entry));
-        errors++;
-    }
-
-    errors += _ql_verify_compress(ql);
-    return errors;
-}
-
-/* Release iterator and verify compress correctly. */
-static void ql_release_iterator(quicklistIter *iter) {
-    quicklist *ql = NULL;
-    if (iter) ql = iter->quicklist;
-    quicklistReleaseIterator(iter);
-    if (ql) assert(!_ql_verify_compress(ql));
-}
-
-/* Generate new string concatenating integer i against string 'prefix' */
-static char *genstr(char *prefix, int i) {
-    static char result[64] = {0};
-    snprintf(result, sizeof(result), "%s%d", prefix, i);
-    return result;
-}
-
-static void randstring(unsigned char *target, size_t sz) {
-    size_t p = 0;
-    int minval, maxval;
-    switch(rand() % 3) {
-    case 0:
-        minval = 'a';
-        maxval = 'z';
-    break;
-    case 1:
-        minval = '0';
-        maxval = '9';
-    break;
-    case 2:
-        minval = 'A';
-        maxval = 'Z';
-    break;
-    default:
-        assert(NULL);
-    }
-
-    while(p < sz)
-        target[p++] = minval+rand()%(maxval-minval+1);
-}
-
-/* main test, but callable from other files */
-int quicklistTest(int argc, char *argv[], int flags) {
-    UNUSED(argc);
-    UNUSED(argv);
-
-    int accurate = (flags & TEST_ACCURATE);
-    unsigned int err = 0;
-    int optimize_start =
-        -(int)(sizeof(optimization_level) / sizeof(*optimization_level));
-
-    printf("Starting optimization offset at: %d\n", optimize_start);
-
-    int options[] = {0, 1, 2, 3, 4, 5, 6, 10};
-    int fills[] = {-5, -4, -3, -2, -1, 0,
-                   1, 2, 32, 66, 128, 999};
-    size_t option_count = sizeof(options) / sizeof(*options);
-    int fill_count = (int)(sizeof(fills) / sizeof(*fills));
-    long long runtime[option_count];
-
-    for (int _i = 0; _i < (int)option_count; _i++) {
-        printf("Testing Compression option %d\n", options[_i]);
-        long long start = mstime();
-        quicklistIter *iter;
-
-        TEST("create list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("add to tail of empty list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistPushTail(ql, "hello", 6);
-            /* 1 for head and 1 for tail because 1 node = head = tail */
-            ql_verify(ql, 1, 1, 1, 1);
-            quicklistRelease(ql);
-        }
-
-        TEST("add to head of empty list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistPushHead(ql, "hello", 6);
-            /* 1 for head and 1 for tail because 1 node = head = tail */
-            ql_verify(ql, 1, 1, 1, 1);
-            quicklistRelease(ql);
-        }
-
-        TEST_DESC("add to tail 5x at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 5; i++)
-                    quicklistPushTail(ql, genstr("hello", i), 32);
-                if (ql->count != 5)
-                    ERROR;
-                if (fills[f] == 32)
-                    ql_verify(ql, 1, 5, 5, 5);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("add to head 5x at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 5; i++)
-                    quicklistPushHead(ql, genstr("hello", i), 32);
-                if (ql->count != 5)
-                    ERROR;
-                if (fills[f] == 32)
-                    ql_verify(ql, 1, 5, 5, 5);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("add to tail 500x at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 500; i++)
-                    quicklistPushTail(ql, genstr("hello", i), 64);
-                if (ql->count != 500)
-                    ERROR;
-                if (fills[f] == 32)
-                    ql_verify(ql, 16, 500, 32, 20);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("add to head 500x at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 500; i++)
-                    quicklistPushHead(ql, genstr("hello", i), 32);
-                if (ql->count != 500)
-                    ERROR;
-                if (fills[f] == 32)
-                    ql_verify(ql, 16, 500, 20, 32);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST("rotate empty") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistRotate(ql);
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("Comprassion Plain node") {
-        for (int f = 0; f < fill_count; f++) {
-            size_t large_limit = (fills[f] < 0) ? quicklistNodeNegFillLimit(fills[f]) + 1 : SIZE_SAFETY_LIMIT + 1;
-
-            char buf[large_limit];
-            quicklist *ql = quicklistNew(fills[f], 1);
-            for (int i = 0; i < 500; i++) {
-                /* Set to 256 to allow the node to be triggered to compress,
-                 * if it is less than 48(nocompress), the test will be successful. */
-                snprintf(buf, sizeof(buf), "hello%d", i);
-                quicklistPushHead(ql, buf, large_limit);
-            }
-
-            quicklistIter *iter = quicklistGetIterator(ql, AL_START_TAIL);
-            quicklistEntry entry;
-            int i = 0;
-            while (quicklistNext(iter, &entry)) {
-                assert(QL_NODE_IS_PLAIN(entry.node));
-                snprintf(buf, sizeof(buf), "hello%d", i);
-                if (strcmp((char *)entry.value, buf))
-                    ERR("value [%s] didn't match [%s] at position %d",
-                        entry.value, buf, i);
-                i++;
-            }
-            ql_release_iterator(iter);
-            quicklistRelease(ql);
-        }
-        }
-
-        TEST("NEXT plain node") {
-        for (int f = 0; f < fill_count; f++) {
-            size_t large_limit = (fills[f] < 0) ? quicklistNodeNegFillLimit(fills[f]) + 1 : SIZE_SAFETY_LIMIT + 1;
-            quicklist *ql = quicklistNew(fills[f], options[_i]);
-
-            char buf[large_limit];
-            memcpy(buf, "plain", 5);
-            quicklistPushHead(ql, buf, large_limit);
-            quicklistPushHead(ql, buf, large_limit);
-            quicklistPushHead(ql, "packed3", 7);
-            quicklistPushHead(ql, "packed4", 7);
-            quicklistPushHead(ql, buf, large_limit);
-
-            quicklistEntry entry;
-            quicklistIter *iter = quicklistGetIterator(ql, AL_START_TAIL);
-
-            while(quicklistNext(iter, &entry) != 0) {
-                if (QL_NODE_IS_PLAIN(entry.node))
-                    assert(!memcmp(entry.value, "plain", 5));
-                else
-                    assert(!memcmp(entry.value, "packed", 6));
-            }
-            ql_release_iterator(iter);
-            quicklistRelease(ql);
-        }
-        }
-
-        TEST("rotate plain node ") {
-        for (int f = 0; f < fill_count; f++) {
-            size_t large_limit = (fills[f] < 0) ? quicklistNodeNegFillLimit(fills[f]) + 1 : SIZE_SAFETY_LIMIT + 1;
-
-            unsigned char *data = NULL;
-            size_t sz;
-            long long lv;
-            int i =0;
-            quicklist *ql = quicklistNew(fills[f], options[_i]);
-            char buf[large_limit];
-            memcpy(buf, "hello1", 6);
-            quicklistPushHead(ql, buf, large_limit);
-            memcpy(buf, "hello4", 6);
-            quicklistPushHead(ql, buf, large_limit);
-            memcpy(buf, "hello3", 6);
-            quicklistPushHead(ql, buf, large_limit);
-            memcpy(buf, "hello2", 6);
-            quicklistPushHead(ql, buf, large_limit);
-            quicklistRotate(ql);
-
-            for(i = 1 ; i < 5; i++) {
-                assert(QL_NODE_IS_PLAIN(ql->tail));
-                quicklistPop(ql, QUICKLIST_HEAD, &data, &sz, &lv);
-                int temp_char = data[5];
-                zfree(data);
-                assert(temp_char == ('0' + i));
-            }
-
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-        }
-
-        TEST("rotate one val once") {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                quicklistPushHead(ql, "hello", 6);
-                quicklistRotate(ql);
-                /* Ignore compression verify because listpack is
-                 * too small to compress. */
-                ql_verify(ql, 1, 1, 1, 1);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("rotate 500 val 5000 times at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                quicklistPushHead(ql, "900", 3);
-                quicklistPushHead(ql, "7000", 4);
-                quicklistPushHead(ql, "-1200", 5);
-                quicklistPushHead(ql, "42", 2);
-                for (int i = 0; i < 500; i++)
-                    quicklistPushHead(ql, genstr("hello", i), 64);
-                ql_info(ql);
-                for (int i = 0; i < 5000; i++) {
-                    ql_info(ql);
-                    quicklistRotate(ql);
-                }
-                if (fills[f] == 1)
-                    ql_verify(ql, 504, 504, 1, 1);
-                else if (fills[f] == 2)
-                    ql_verify(ql, 252, 504, 2, 2);
-                else if (fills[f] == 32)
-                    ql_verify(ql, 16, 504, 32, 24);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST("pop empty") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistPop(ql, QUICKLIST_HEAD, NULL, NULL, NULL);
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("pop 1 string from 1") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            char *populate = genstr("hello", 331);
-            quicklistPushHead(ql, populate, 32);
-            unsigned char *data;
-            size_t sz;
-            long long lv;
-            ql_info(ql);
-            assert(quicklistPop(ql, QUICKLIST_HEAD, &data, &sz, &lv));
-            assert(data != NULL);
-            assert(sz == 32);
-            if (strcmp(populate, (char *)data)) {
-                int size = sz;
-                ERR("Pop'd value (%.*s) didn't equal original value (%s)", size,
-                    data, populate);
-            }
-            zfree(data);
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("pop head 1 number from 1") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistPushHead(ql, "55513", 5);
-            unsigned char *data;
-            size_t sz;
-            long long lv;
-            ql_info(ql);
-            assert(quicklistPop(ql, QUICKLIST_HEAD, &data, &sz, &lv));
-            assert(data == NULL);
-            assert(lv == 55513);
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("pop head 500 from 500") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            for (int i = 0; i < 500; i++)
-                quicklistPushHead(ql, genstr("hello", i), 32);
-            ql_info(ql);
-            for (int i = 0; i < 500; i++) {
-                unsigned char *data;
-                size_t sz;
-                long long lv;
-                int ret = quicklistPop(ql, QUICKLIST_HEAD, &data, &sz, &lv);
-                assert(ret == 1);
-                assert(data != NULL);
-                assert(sz == 32);
-                if (strcmp(genstr("hello", 499 - i), (char *)data)) {
-                    int size = sz;
-                    ERR("Pop'd value (%.*s) didn't equal original value (%s)",
-                        size, data, genstr("hello", 499 - i));
-                }
-                zfree(data);
-            }
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("pop head 5000 from 500") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            for (int i = 0; i < 500; i++)
-                quicklistPushHead(ql, genstr("hello", i), 32);
-            for (int i = 0; i < 5000; i++) {
-                unsigned char *data;
-                size_t sz;
-                long long lv;
-                int ret = quicklistPop(ql, QUICKLIST_HEAD, &data, &sz, &lv);
-                if (i < 500) {
-                    assert(ret == 1);
-                    assert(data != NULL);
-                    assert(sz == 32);
-                    if (strcmp(genstr("hello", 499 - i), (char *)data)) {
-                        int size = sz;
-                        ERR("Pop'd value (%.*s) didn't equal original value "
-                            "(%s)",
-                            size, data, genstr("hello", 499 - i));
-                    }
-                    zfree(data);
-                } else {
-                    assert(ret == 0);
-                }
-            }
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("iterate forward over 500 list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            for (int i = 0; i < 500; i++)
-                quicklistPushHead(ql, genstr("hello", i), 32);
-            quicklistIter *iter = quicklistGetIterator(ql, AL_START_HEAD);
-            quicklistEntry entry;
-            int i = 499, count = 0;
-            while (quicklistNext(iter, &entry)) {
-                char *h = genstr("hello", i);
-                if (strcmp((char *)entry.value, h))
-                    ERR("value [%s] didn't match [%s] at position %d",
-                        entry.value, h, i);
-                i--;
-                count++;
-            }
-            if (count != 500)
-                ERR("Didn't iterate over exactly 500 elements (%d)", i);
-            ql_verify(ql, 16, 500, 20, 32);
-            ql_release_iterator(iter);
-            quicklistRelease(ql);
-        }
-
-        TEST("iterate reverse over 500 list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            for (int i = 0; i < 500; i++)
-                quicklistPushHead(ql, genstr("hello", i), 32);
-            quicklistIter *iter = quicklistGetIterator(ql, AL_START_TAIL);
-            quicklistEntry entry;
-            int i = 0;
-            while (quicklistNext(iter, &entry)) {
-                char *h = genstr("hello", i);
-                if (strcmp((char *)entry.value, h))
-                    ERR("value [%s] didn't match [%s] at position %d",
-                        entry.value, h, i);
-                i++;
-            }
-            if (i != 500)
-                ERR("Didn't iterate over exactly 500 elements (%d)", i);
-            ql_verify(ql, 16, 500, 20, 32);
-            ql_release_iterator(iter);
-            quicklistRelease(ql);
-        }
-
-        TEST("insert after 1 element") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistPushHead(ql, "hello", 6);
-            quicklistEntry entry;
-            iter = quicklistGetIteratorEntryAtIdx(ql, 0, &entry);
-            quicklistInsertAfter(iter, &entry, "abc", 4);
-            ql_release_iterator(iter);
-            ql_verify(ql, 1, 2, 2, 2);
-
-            /* verify results */
-            iter = quicklistGetIteratorEntryAtIdx(ql, 0, &entry);
-            int sz = entry.sz;
-            if (strncmp((char *)entry.value, "hello", 5)) {
-                ERR("Value 0 didn't match, instead got: %.*s", sz,
-                    entry.value);
-            }
-            ql_release_iterator(iter);
-
-            iter = quicklistGetIteratorEntryAtIdx(ql, 1, &entry);
-            sz = entry.sz;
-            if (strncmp((char *)entry.value, "abc", 3)) {
-                ERR("Value 1 didn't match, instead got: %.*s", sz,
-                    entry.value);
-            }
-            ql_release_iterator(iter);
-            quicklistRelease(ql);
-        }
-
-        TEST("insert before 1 element") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistPushHead(ql, "hello", 6);
-            quicklistEntry entry;
-            iter = quicklistGetIteratorEntryAtIdx(ql, 0, &entry);
-            quicklistInsertBefore(iter, &entry, "abc", 4);
-            ql_release_iterator(iter);
-            ql_verify(ql, 1, 2, 2, 2);
-
-            /* verify results */
-            iter = quicklistGetIteratorEntryAtIdx(ql, 0, &entry);
-            int sz = entry.sz;
-            if (strncmp((char *)entry.value, "abc", 3)) {
-                ERR("Value 0 didn't match, instead got: %.*s", sz,
-                    entry.value);
-            }
-            ql_release_iterator(iter);
-
-            iter = quicklistGetIteratorEntryAtIdx(ql, 1, &entry);
-            sz = entry.sz;
-            if (strncmp((char *)entry.value, "hello", 5)) {
-                ERR("Value 1 didn't match, instead got: %.*s", sz,
-                    entry.value);
-            }
-            ql_release_iterator(iter);
-            quicklistRelease(ql);
-        }
-
-        TEST("insert head while head node is full") {
-            quicklist *ql = quicklistNew(4, options[_i]);
-            for (int i = 0; i < 10; i++)
-                quicklistPushTail(ql, genstr("hello", i), 6);
-            quicklistSetFill(ql, -1);
-            quicklistEntry entry;
-            iter = quicklistGetIteratorEntryAtIdx(ql, -10, &entry);
-            char buf[4096] = {0};
-            quicklistInsertBefore(iter, &entry, buf, 4096);
-            ql_release_iterator(iter);
-            ql_verify(ql, 4, 11, 1, 2);
-            quicklistRelease(ql);
-        }
-
-        TEST("insert tail while tail node is full") {
-            quicklist *ql = quicklistNew(4, options[_i]);
-            for (int i = 0; i < 10; i++)
-                quicklistPushHead(ql, genstr("hello", i), 6);
-            quicklistSetFill(ql, -1);
-            quicklistEntry entry;
-            iter = quicklistGetIteratorEntryAtIdx(ql, -1, &entry);
-            char buf[4096] = {0};
-            quicklistInsertAfter(iter, &entry, buf, 4096);
-            ql_release_iterator(iter);
-            ql_verify(ql, 4, 11, 2, 1);
-            quicklistRelease(ql);
-        }
-
-        TEST_DESC("insert once in elements while iterating at compress %d",
-                  options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                quicklistPushTail(ql, "abc", 3);
-                quicklistSetFill(ql, 1);
-                quicklistPushTail(ql, "def", 3); /* force to unique node */
-                quicklistSetFill(ql, f);
-                quicklistPushTail(ql, "bob", 3); /* force to reset for +3 */
-                quicklistPushTail(ql, "foo", 3);
-                quicklistPushTail(ql, "zoo", 3);
-
-                itrprintr(ql, 0);
-                /* insert "bar" before "bob" while iterating over list. */
-                quicklistIter *iter = quicklistGetIterator(ql, AL_START_HEAD);
-                quicklistEntry entry;
-                while (quicklistNext(iter, &entry)) {
-                    if (!strncmp((char *)entry.value, "bob", 3)) {
-                        /* Insert as fill = 1 so it spills into new node. */
-                        quicklistInsertBefore(iter, &entry, "bar", 3);
-                        break; /* didn't we fix insert-while-iterating? */
-                    }
-                }
-                ql_release_iterator(iter);
-                itrprintr(ql, 0);
-
-                /* verify results */
-                iter = quicklistGetIteratorEntryAtIdx(ql, 0, &entry);
-                int sz = entry.sz;
-
-                if (strncmp((char *)entry.value, "abc", 3))
-                    ERR("Value 0 didn't match, instead got: %.*s", sz,
-                        entry.value);
-                ql_release_iterator(iter);
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, 1, &entry);
-                if (strncmp((char *)entry.value, "def", 3))
-                    ERR("Value 1 didn't match, instead got: %.*s", sz,
-                        entry.value);
-                ql_release_iterator(iter);
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, 2, &entry);
-                if (strncmp((char *)entry.value, "bar", 3))
-                    ERR("Value 2 didn't match, instead got: %.*s", sz,
-                        entry.value);
-                ql_release_iterator(iter);
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, 3, &entry);
-                if (strncmp((char *)entry.value, "bob", 3))
-                    ERR("Value 3 didn't match, instead got: %.*s", sz,
-                        entry.value);
-                ql_release_iterator(iter);
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, 4, &entry);
-                if (strncmp((char *)entry.value, "foo", 3))
-                    ERR("Value 4 didn't match, instead got: %.*s", sz,
-                        entry.value);
-                ql_release_iterator(iter);
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, 5, &entry);
-                if (strncmp((char *)entry.value, "zoo", 3))
-                    ERR("Value 5 didn't match, instead got: %.*s", sz,
-                        entry.value);
-                ql_release_iterator(iter);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("insert [before] 250 new in middle of 500 elements at compress %d",
-                  options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 500; i++)
-                    quicklistPushTail(ql, genstr("hello", i), 32);
-                for (int i = 0; i < 250; i++) {
-                    quicklistEntry entry;
-                    iter = quicklistGetIteratorEntryAtIdx(ql, 250, &entry);
-                    quicklistInsertBefore(iter, &entry, genstr("abc", i), 32);
-                    ql_release_iterator(iter);
-                }
-                if (fills[f] == 32)
-                    ql_verify(ql, 25, 750, 32, 20);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("insert [after] 250 new in middle of 500 elements at compress %d",
-                  options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 500; i++)
-                    quicklistPushHead(ql, genstr("hello", i), 32);
-                for (int i = 0; i < 250; i++) {
-                    quicklistEntry entry;
-                    iter = quicklistGetIteratorEntryAtIdx(ql, 250, &entry);
-                    quicklistInsertAfter(iter, &entry, genstr("abc", i), 32);
-                    ql_release_iterator(iter);
-                }
-
-                if (ql->count != 750)
-                    ERR("List size not 750, but rather %ld", ql->count);
-
-                if (fills[f] == 32)
-                    ql_verify(ql, 26, 750, 20, 32);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST("duplicate empty list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklist *copy = quicklistDup(ql);
-            ql_verify(copy, 0, 0, 0, 0);
-            quicklistRelease(ql);
-            quicklistRelease(copy);
-        }
-
-        TEST("duplicate list of 1 element") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistPushHead(ql, genstr("hello", 3), 32);
-            ql_verify(ql, 1, 1, 1, 1);
-            quicklist *copy = quicklistDup(ql);
-            ql_verify(copy, 1, 1, 1, 1);
-            quicklistRelease(ql);
-            quicklistRelease(copy);
-        }
-
-        TEST("duplicate list of 500") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            for (int i = 0; i < 500; i++)
-                quicklistPushHead(ql, genstr("hello", i), 32);
-            ql_verify(ql, 16, 500, 20, 32);
-
-            quicklist *copy = quicklistDup(ql);
-            ql_verify(copy, 16, 500, 20, 32);
-            quicklistRelease(ql);
-            quicklistRelease(copy);
-        }
-
-        for (int f = 0; f < fill_count; f++) {
-            TEST_DESC("index 1,200 from 500 list at fill %d at compress %d", f,
-                      options[_i]) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 500; i++)
-                    quicklistPushTail(ql, genstr("hello", i + 1), 32);
-                quicklistEntry entry;
-                iter = quicklistGetIteratorEntryAtIdx(ql, 1, &entry);
-                if (strcmp((char *)entry.value, "hello2") != 0)
-                    ERR("Value: %s", entry.value);
-                ql_release_iterator(iter);
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, 200, &entry);
-                if (strcmp((char *)entry.value, "hello201") != 0)
-                    ERR("Value: %s", entry.value);
-                ql_release_iterator(iter);
-                quicklistRelease(ql);
-            }
-
-            TEST_DESC("index -1,-2 from 500 list at fill %d at compress %d",
-                      fills[f], options[_i]) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 500; i++)
-                    quicklistPushTail(ql, genstr("hello", i + 1), 32);
-                quicklistEntry entry;
-                iter = quicklistGetIteratorEntryAtIdx(ql, -1, &entry);
-                if (strcmp((char *)entry.value, "hello500") != 0)
-                    ERR("Value: %s", entry.value);
-                ql_release_iterator(iter);
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, -2, &entry);
-                if (strcmp((char *)entry.value, "hello499") != 0)
-                    ERR("Value: %s", entry.value);
-                ql_release_iterator(iter);
-                quicklistRelease(ql);
-            }
-
-            TEST_DESC("index -100 from 500 list at fill %d at compress %d",
-                      fills[f], options[_i]) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 500; i++)
-                    quicklistPushTail(ql, genstr("hello", i + 1), 32);
-                quicklistEntry entry;
-                iter = quicklistGetIteratorEntryAtIdx(ql, -100, &entry);
-                if (strcmp((char *)entry.value, "hello401") != 0)
-                    ERR("Value: %s", entry.value);
-                ql_release_iterator(iter);
-                quicklistRelease(ql);
-            }
-
-            TEST_DESC("index too big +1 from 50 list at fill %d at compress %d",
-                      fills[f], options[_i]) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                for (int i = 0; i < 50; i++)
-                    quicklistPushTail(ql, genstr("hello", i + 1), 32);
-                quicklistEntry entry;
-                int sz = entry.sz;
-                iter = quicklistGetIteratorEntryAtIdx(ql, 50, &entry);
-                if (iter)
-                    ERR("Index found at 50 with 50 list: %.*s", sz,
-                        entry.value);
-                ql_release_iterator(iter);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST("delete range empty list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistDelRange(ql, 5, 20);
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("delete range of entire node in list of one node") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            for (int i = 0; i < 32; i++)
-                quicklistPushHead(ql, genstr("hello", i), 32);
-            ql_verify(ql, 1, 32, 32, 32);
-            quicklistDelRange(ql, 0, 32);
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("delete range of entire node with overflow counts") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            for (int i = 0; i < 32; i++)
-                quicklistPushHead(ql, genstr("hello", i), 32);
-            ql_verify(ql, 1, 32, 32, 32);
-            quicklistDelRange(ql, 0, 128);
-            ql_verify(ql, 0, 0, 0, 0);
-            quicklistRelease(ql);
-        }
-
-        TEST("delete middle 100 of 500 list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            for (int i = 0; i < 500; i++)
-                quicklistPushTail(ql, genstr("hello", i + 1), 32);
-            ql_verify(ql, 16, 500, 32, 20);
-            quicklistDelRange(ql, 200, 100);
-            ql_verify(ql, 14, 400, 32, 20);
-            quicklistRelease(ql);
-        }
-
-        TEST("delete less than fill but across nodes") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            for (int i = 0; i < 500; i++)
-                quicklistPushTail(ql, genstr("hello", i + 1), 32);
-            ql_verify(ql, 16, 500, 32, 20);
-            quicklistDelRange(ql, 60, 10);
-            ql_verify(ql, 16, 490, 32, 20);
-            quicklistRelease(ql);
-        }
-
-        TEST("delete negative 1 from 500 list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            for (int i = 0; i < 500; i++)
-                quicklistPushTail(ql, genstr("hello", i + 1), 32);
-            ql_verify(ql, 16, 500, 32, 20);
-            quicklistDelRange(ql, -1, 1);
-            ql_verify(ql, 16, 499, 32, 19);
-            quicklistRelease(ql);
-        }
-
-        TEST("delete negative 1 from 500 list with overflow counts") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            for (int i = 0; i < 500; i++)
-                quicklistPushTail(ql, genstr("hello", i + 1), 32);
-            ql_verify(ql, 16, 500, 32, 20);
-            quicklistDelRange(ql, -1, 128);
-            ql_verify(ql, 16, 499, 32, 19);
-            quicklistRelease(ql);
-        }
-
-        TEST("delete negative 100 from 500 list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            for (int i = 0; i < 500; i++)
-                quicklistPushTail(ql, genstr("hello", i + 1), 32);
-            quicklistDelRange(ql, -100, 100);
-            ql_verify(ql, 13, 400, 32, 16);
-            quicklistRelease(ql);
-        }
-
-        TEST("delete -10 count 5 from 50 list") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            for (int i = 0; i < 50; i++)
-                quicklistPushTail(ql, genstr("hello", i + 1), 32);
-            ql_verify(ql, 2, 50, 32, 18);
-            quicklistDelRange(ql, -10, 5);
-            ql_verify(ql, 2, 45, 32, 13);
-            quicklistRelease(ql);
-        }
-
-        TEST("numbers only list read") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistPushTail(ql, "1111", 4);
-            quicklistPushTail(ql, "2222", 4);
-            quicklistPushTail(ql, "3333", 4);
-            quicklistPushTail(ql, "4444", 4);
-            ql_verify(ql, 1, 4, 4, 4);
-            quicklistEntry entry;
-            iter = quicklistGetIteratorEntryAtIdx(ql, 0, &entry);
-            if (entry.longval != 1111)
-                ERR("Not 1111, %lld", entry.longval);
-            ql_release_iterator(iter);
-
-            iter = quicklistGetIteratorEntryAtIdx(ql, 1, &entry);
-            if (entry.longval != 2222)
-                ERR("Not 2222, %lld", entry.longval);
-            ql_release_iterator(iter);
-
-            iter = quicklistGetIteratorEntryAtIdx(ql, 2, &entry);
-            if (entry.longval != 3333)
-                ERR("Not 3333, %lld", entry.longval);
-            ql_release_iterator(iter);
-
-            iter = quicklistGetIteratorEntryAtIdx(ql, 3, &entry);
-            if (entry.longval != 4444)
-                ERR("Not 4444, %lld", entry.longval);
-            ql_release_iterator(iter);
-
-            iter = quicklistGetIteratorEntryAtIdx(ql, 4, &entry);
-            if (iter)
-                ERR("Index past elements: %lld", entry.longval);
-            ql_release_iterator(iter);
-            
-            iter = quicklistGetIteratorEntryAtIdx(ql, -1, &entry);
-            if (entry.longval != 4444)
-                ERR("Not 4444 (reverse), %lld", entry.longval);
-            ql_release_iterator(iter);
-
-            iter = quicklistGetIteratorEntryAtIdx(ql, -2, &entry);
-            if (entry.longval != 3333)
-                ERR("Not 3333 (reverse), %lld", entry.longval);
-            ql_release_iterator(iter);
-
-            iter = quicklistGetIteratorEntryAtIdx(ql, -3, &entry);
-            if (entry.longval != 2222)
-                ERR("Not 2222 (reverse), %lld", entry.longval);
-            ql_release_iterator(iter);
-            
-            iter = quicklistGetIteratorEntryAtIdx(ql, -4, &entry);
-            if (entry.longval != 1111)
-                ERR("Not 1111 (reverse), %lld", entry.longval);
-            ql_release_iterator(iter);
-            
-            iter = quicklistGetIteratorEntryAtIdx(ql, -5, &entry);
-            if (iter)
-                ERR("Index past elements (reverse), %lld", entry.longval);
-            ql_release_iterator(iter);
-            quicklistRelease(ql);
-        }
-
-        TEST("numbers larger list read") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistSetFill(ql, 32);
-            char num[32];
-            long long nums[5000];
-            for (int i = 0; i < 5000; i++) {
-                nums[i] = -5157318210846258176 + i;
-                int sz = ll2string(num, sizeof(num), nums[i]);
-                quicklistPushTail(ql, num, sz);
-            }
-            quicklistPushTail(ql, "xxxxxxxxxxxxxxxxxxxx", 20);
-            quicklistEntry entry;
-            for (int i = 0; i < 5000; i++) {
-                iter = quicklistGetIteratorEntryAtIdx(ql, i, &entry);
-                if (entry.longval != nums[i])
-                    ERR("[%d] Not longval %lld but rather %lld", i, nums[i],
-                        entry.longval);
-                entry.longval = 0xdeadbeef;
-                ql_release_iterator(iter);
-            }
-            iter = quicklistGetIteratorEntryAtIdx(ql, 5000, &entry);
-            if (strncmp((char *)entry.value, "xxxxxxxxxxxxxxxxxxxx", 20))
-                ERR("String val not match: %s", entry.value);
-            ql_verify(ql, 157, 5001, 32, 9);
-            ql_release_iterator(iter);
-            quicklistRelease(ql);
-        }
-
-        TEST("numbers larger list read B") {
-            quicklist *ql = quicklistNew(-2, options[_i]);
-            quicklistPushTail(ql, "99", 2);
-            quicklistPushTail(ql, "98", 2);
-            quicklistPushTail(ql, "xxxxxxxxxxxxxxxxxxxx", 20);
-            quicklistPushTail(ql, "96", 2);
-            quicklistPushTail(ql, "95", 2);
-            quicklistReplaceAtIndex(ql, 1, "foo", 3);
-            quicklistReplaceAtIndex(ql, -1, "bar", 3);
-            quicklistRelease(ql);
-        }
-
-        TEST_DESC("lrem test at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                char *words[] = {"abc", "foo", "bar",  "foobar", "foobared",
-                                 "zap", "bar", "test", "foo"};
-                char *result[] = {"abc", "foo",  "foobar", "foobared",
-                                  "zap", "test", "foo"};
-                char *resultB[] = {"abc",      "foo", "foobar",
-                                   "foobared", "zap", "test"};
-                for (int i = 0; i < 9; i++)
-                    quicklistPushTail(ql, words[i], strlen(words[i]));
-
-                /* lrem 0 bar */
-                quicklistIter *iter = quicklistGetIterator(ql, AL_START_HEAD);
-                quicklistEntry entry;
-                int i = 0;
-                while (quicklistNext(iter, &entry)) {
-                    if (quicklistCompare(&entry, (unsigned char *)"bar", 3)) {
-                        quicklistDelEntry(iter, &entry);
-                    }
-                    i++;
-                }
-                ql_release_iterator(iter);
-
-                /* check result of lrem 0 bar */
-                iter = quicklistGetIterator(ql, AL_START_HEAD);
-                i = 0;
-                while (quicklistNext(iter, &entry)) {
-                    /* Result must be: abc, foo, foobar, foobared, zap, test,
-                     * foo */
-                    int sz = entry.sz;
-                    if (strncmp((char *)entry.value, result[i], entry.sz)) {
-                        ERR("No match at position %d, got %.*s instead of %s",
-                            i, sz, entry.value, result[i]);
-                    }
-                    i++;
-                }
-                ql_release_iterator(iter);
-
-                quicklistPushTail(ql, "foo", 3);
-
-                /* lrem -2 foo */
-                iter = quicklistGetIterator(ql, AL_START_TAIL);
-                i = 0;
-                int del = 2;
-                while (quicklistNext(iter, &entry)) {
-                    if (quicklistCompare(&entry, (unsigned char *)"foo", 3)) {
-                        quicklistDelEntry(iter, &entry);
-                        del--;
-                    }
-                    if (!del)
-                        break;
-                    i++;
-                }
-                ql_release_iterator(iter);
-
-                /* check result of lrem -2 foo */
-                /* (we're ignoring the '2' part and still deleting all foo
-                 * because
-                 * we only have two foo) */
-                iter = quicklistGetIterator(ql, AL_START_TAIL);
-                i = 0;
-                size_t resB = sizeof(resultB) / sizeof(*resultB);
-                while (quicklistNext(iter, &entry)) {
-                    /* Result must be: abc, foo, foobar, foobared, zap, test,
-                     * foo */
-                    int sz = entry.sz;
-                    if (strncmp((char *)entry.value, resultB[resB - 1 - i],
-                                sz)) {
-                        ERR("No match at position %d, got %.*s instead of %s",
-                            i, sz, entry.value, resultB[resB - 1 - i]);
-                    }
-                    i++;
-                }
-
-                ql_release_iterator(iter);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("iterate reverse + delete at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                quicklistPushTail(ql, "abc", 3);
-                quicklistPushTail(ql, "def", 3);
-                quicklistPushTail(ql, "hij", 3);
-                quicklistPushTail(ql, "jkl", 3);
-                quicklistPushTail(ql, "oop", 3);
-
-                quicklistEntry entry;
-                quicklistIter *iter = quicklistGetIterator(ql, AL_START_TAIL);
-                int i = 0;
-                while (quicklistNext(iter, &entry)) {
-                    if (quicklistCompare(&entry, (unsigned char *)"hij", 3)) {
-                        quicklistDelEntry(iter, &entry);
-                    }
-                    i++;
-                }
-                ql_release_iterator(iter);
-
-                if (i != 5)
-                    ERR("Didn't iterate 5 times, iterated %d times.", i);
-
-                /* Check results after deletion of "hij" */
-                iter = quicklistGetIterator(ql, AL_START_HEAD);
-                i = 0;
-                char *vals[] = {"abc", "def", "jkl", "oop"};
-                while (quicklistNext(iter, &entry)) {
-                    if (!quicklistCompare(&entry, (unsigned char *)vals[i],
-                                          3)) {
-                        ERR("Value at %d didn't match %s\n", i, vals[i]);
-                    }
-                    i++;
-                }
-                ql_release_iterator(iter);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("iterator at index test at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                char num[32];
-                long long nums[5000];
-                for (int i = 0; i < 760; i++) {
-                    nums[i] = -5157318210846258176 + i;
-                    int sz = ll2string(num, sizeof(num), nums[i]);
-                    quicklistPushTail(ql, num, sz);
-                }
-
-                quicklistEntry entry;
-                quicklistIter *iter =
-                    quicklistGetIteratorAtIdx(ql, AL_START_HEAD, 437);
-                int i = 437;
-                while (quicklistNext(iter, &entry)) {
-                    if (entry.longval != nums[i])
-                        ERR("Expected %lld, but got %lld", entry.longval,
-                            nums[i]);
-                    i++;
-                }
-                ql_release_iterator(iter);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("ltrim test A at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                char num[32];
-                long long nums[5000];
-                for (int i = 0; i < 32; i++) {
-                    nums[i] = -5157318210846258176 + i;
-                    int sz = ll2string(num, sizeof(num), nums[i]);
-                    quicklistPushTail(ql, num, sz);
-                }
-                if (fills[f] == 32)
-                    ql_verify(ql, 1, 32, 32, 32);
-                /* ltrim 25 53 (keep [25,32] inclusive = 7 remaining) */
-                quicklistDelRange(ql, 0, 25);
-                quicklistDelRange(ql, 0, 0);
-                quicklistEntry entry;
-                for (int i = 0; i < 7; i++) {
-                    iter = quicklistGetIteratorEntryAtIdx(ql, i, &entry);
-                    if (entry.longval != nums[25 + i])
-                        ERR("Deleted invalid range!  Expected %lld but got "
-                            "%lld",
-                            entry.longval, nums[25 + i]);
-                    ql_release_iterator(iter);
-                }
-                if (fills[f] == 32)
-                    ql_verify(ql, 1, 7, 7, 7);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("ltrim test B at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                /* Force-disable compression because our 33 sequential
-                 * integers don't compress and the check always fails. */
-                quicklist *ql = quicklistNew(fills[f], QUICKLIST_NOCOMPRESS);
-                char num[32];
-                long long nums[5000];
-                for (int i = 0; i < 33; i++) {
-                    nums[i] = i;
-                    int sz = ll2string(num, sizeof(num), nums[i]);
-                    quicklistPushTail(ql, num, sz);
-                }
-                if (fills[f] == 32)
-                    ql_verify(ql, 2, 33, 32, 1);
-                /* ltrim 5 16 (keep [5,16] inclusive = 12 remaining) */
-                quicklistDelRange(ql, 0, 5);
-                quicklistDelRange(ql, -16, 16);
-                if (fills[f] == 32)
-                    ql_verify(ql, 1, 12, 12, 12);
-                quicklistEntry entry;
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, 0, &entry);
-                if (entry.longval != 5)
-                    ERR("A: longval not 5, but %lld", entry.longval);
-                ql_release_iterator(iter);
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, -1, &entry);
-                if (entry.longval != 16)
-                    ERR("B! got instead: %lld", entry.longval);
-                quicklistPushTail(ql, "bobobob", 7);
-                ql_release_iterator(iter);
-
-                iter = quicklistGetIteratorEntryAtIdx(ql, -1, &entry);
-                int sz = entry.sz;
-                if (strncmp((char *)entry.value, "bobobob", 7))
-                    ERR("Tail doesn't match bobobob, it's %.*s instead",
-                        sz, entry.value);
-                ql_release_iterator(iter);
-
-                for (int i = 0; i < 12; i++) {
-                    iter = quicklistGetIteratorEntryAtIdx(ql, i, &entry);
-                    if (entry.longval != nums[5 + i])
-                        ERR("Deleted invalid range!  Expected %lld but got "
-                            "%lld",
-                            entry.longval, nums[5 + i]);
-                    ql_release_iterator(iter);
-                }
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("ltrim test C at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                char num[32];
-                long long nums[5000];
-                for (int i = 0; i < 33; i++) {
-                    nums[i] = -5157318210846258176 + i;
-                    int sz = ll2string(num, sizeof(num), nums[i]);
-                    quicklistPushTail(ql, num, sz);
-                }
-                if (fills[f] == 32)
-                    ql_verify(ql, 2, 33, 32, 1);
-                /* ltrim 3 3 (keep [3,3] inclusive = 1 remaining) */
-                quicklistDelRange(ql, 0, 3);
-                quicklistDelRange(ql, -29,
-                                  4000); /* make sure not loop forever */
-                if (fills[f] == 32)
-                    ql_verify(ql, 1, 1, 1, 1);
-                quicklistEntry entry;
-                iter = quicklistGetIteratorEntryAtIdx(ql, 0, &entry);
-                if (entry.longval != -5157318210846258173)
-                    ERROR;
-                ql_release_iterator(iter);
-                quicklistRelease(ql);
-            }
-        }
-
-        TEST_DESC("ltrim test D at compress %d", options[_i]) {
-            for (int f = 0; f < fill_count; f++) {
-                quicklist *ql = quicklistNew(fills[f], options[_i]);
-                char num[32];
-                long long nums[5000];
-                for (int i = 0; i < 33; i++) {
-                    nums[i] = -5157318210846258176 + i;
-                    int sz = ll2string(num, sizeof(num), nums[i]);
-                    quicklistPushTail(ql, num, sz);
-                }
-                if (fills[f] == 32)
-                    ql_verify(ql, 2, 33, 32, 1);
-                quicklistDelRange(ql, -12, 3);
-                if (ql->count != 30)
-                    ERR("Didn't delete exactly three elements!  Count is: %lu",
-                        ql->count);
-                quicklistRelease(ql);
-            }
-        }
-
-        long long stop = mstime();
-        runtime[_i] = stop - start;
-    }
-
-    /* Run a longer test of compression depth outside of primary test loop. */
-    int list_sizes[] = {250, 251, 500, 999, 1000};
-    long long start = mstime();
-    int list_count = accurate ? (int)(sizeof(list_sizes) / sizeof(*list_sizes)) : 1;
-    for (int list = 0; list < list_count; list++) {
-        TEST_DESC("verify specific compression of interior nodes with %d list ",
-                  list_sizes[list]) {
-            for (int f = 0; f < fill_count; f++) {
-                for (int depth = 1; depth < 40; depth++) {
-                    /* skip over many redundant test cases */
-                    quicklist *ql = quicklistNew(fills[f], depth);
-                    for (int i = 0; i < list_sizes[list]; i++) {
-                        quicklistPushTail(ql, genstr("hello TAIL", i + 1), 64);
-                        quicklistPushHead(ql, genstr("hello HEAD", i + 1), 64);
-                    }
-
-                    for (int step = 0; step < 2; step++) {
-                        /* test remove node */
-                        if (step == 1) {
-                            for (int i = 0; i < list_sizes[list] / 2; i++) {
-                                unsigned char *data;
-                                assert(quicklistPop(ql, QUICKLIST_HEAD, &data,
-                                                    NULL, NULL));
-                                zfree(data);
-                                assert(quicklistPop(ql, QUICKLIST_TAIL, &data,
-                                                    NULL, NULL));
-                                zfree(data);
-                            }
-                        }
-                        quicklistNode *node = ql->head;
-                        unsigned int low_raw = ql->compress;
-                        unsigned int high_raw = ql->len - ql->compress;
-
-                        for (unsigned int at = 0; at < ql->len;
-                            at++, node = node->next) {
-                            if (at < low_raw || at >= high_raw) {
-                                if (node->encoding != QUICKLIST_NODE_ENCODING_RAW) {
-                                    ERR("Incorrect compression: node %d is "
-                                        "compressed at depth %d ((%u, %u); total "
-                                        "nodes: %lu; size: %zu)",
-                                        at, depth, low_raw, high_raw, ql->len,
-                                        node->sz);
-                                }
-                            } else {
-                                if (node->encoding != QUICKLIST_NODE_ENCODING_LZF) {
-                                    ERR("Incorrect non-compression: node %d is NOT "
-                                        "compressed at depth %d ((%u, %u); total "
-                                        "nodes: %lu; size: %zu; attempted: %d)",
-                                        at, depth, low_raw, high_raw, ql->len,
-                                        node->sz, node->attempted_compress);
-                                }
-                            }
-                        }
-                    }
-
-                    quicklistRelease(ql);
-                }
-            }
-        }
-    }
-    long long stop = mstime();
-
-    printf("\n");
-    for (size_t i = 0; i < option_count; i++)
-        printf("Test Loop %02d: %0.2f seconds.\n", options[i],
-               (float)runtime[i] / 1000);
-    printf("Compressions: %0.2f seconds.\n", (float)(stop - start) / 1000);
-    printf("\n");
-
-    TEST("bookmark get updated to next item") {
-        quicklist *ql = quicklistNew(1, 0);
-        quicklistPushTail(ql, "1", 1);
-        quicklistPushTail(ql, "2", 1);
-        quicklistPushTail(ql, "3", 1);
-        quicklistPushTail(ql, "4", 1);
-        quicklistPushTail(ql, "5", 1);
-        assert(ql->len==5);
-        /* add two bookmarks, one pointing to the node before the last. */
-        assert(quicklistBookmarkCreate(&ql, "_dummy", ql->head->next));
-        assert(quicklistBookmarkCreate(&ql, "_test", ql->tail->prev));
-        /* test that the bookmark returns the right node, delete it and see that the bookmark points to the last node */
-        assert(quicklistBookmarkFind(ql, "_test") == ql->tail->prev);
-        assert(quicklistDelRange(ql, -2, 1));
-        assert(quicklistBookmarkFind(ql, "_test") == ql->tail);
-        /* delete the last node, and see that the bookmark was deleted. */
-        assert(quicklistDelRange(ql, -1, 1));
-        assert(quicklistBookmarkFind(ql, "_test") == NULL);
-        /* test that other bookmarks aren't affected */
-        assert(quicklistBookmarkFind(ql, "_dummy") == ql->head->next);
-        assert(quicklistBookmarkFind(ql, "_missing") == NULL);
-        assert(ql->len==3);
-        quicklistBookmarksClear(ql); /* for coverage */
-        assert(quicklistBookmarkFind(ql, "_dummy") == NULL);
-        quicklistRelease(ql);
-    }
-
-    TEST("bookmark limit") {
-        int i;
-        quicklist *ql = quicklistNew(1, 0);
-        quicklistPushHead(ql, "1", 1);
-        for (i=0; i<QL_MAX_BM; i++)
-            assert(quicklistBookmarkCreate(&ql, genstr("",i), ql->head));
-        /* when all bookmarks are used, creation fails */
-        assert(!quicklistBookmarkCreate(&ql, "_test", ql->head));
-        /* delete one and see that we can now create another */
-        assert(quicklistBookmarkDelete(ql, "0"));
-        assert(quicklistBookmarkCreate(&ql, "_test", ql->head));
-        /* delete one and see that the rest survive */
-        assert(quicklistBookmarkDelete(ql, "_test"));
-        for (i=1; i<QL_MAX_BM; i++)
-            assert(quicklistBookmarkFind(ql, genstr("",i)) == ql->head);
-        /* make sure the deleted ones are indeed gone */
-        assert(!quicklistBookmarkFind(ql, "0"));
-        assert(!quicklistBookmarkFind(ql, "_test"));
-        quicklistRelease(ql);
-    }
-
-    if (flags & TEST_LARGE_MEMORY) {
-        TEST("compress and decompress quicklist listpack node") {
-            quicklistNode *node = quicklistCreateNode();
-            node->entry = lpNew(0);
-
-            /* Just to avoid triggering the assertion in __quicklistCompressNode(),
-             * it disables the passing of quicklist head or tail node. */
-            node->prev = quicklistCreateNode();
-            node->next = quicklistCreateNode();
-            
-            /* Create a rand string */
-            size_t sz = (1 << 25); /* 32MB per one entry */
-            unsigned char *s = zmalloc(sz);
-            randstring(s, sz);
-
-            /* Keep filling the node, until it reaches 1GB */
-            for (int i = 0; i < 32; i++) {
-                node->entry = lpAppend(node->entry, s, sz);
-                quicklistNodeUpdateSz(node);
-
-                long long start = mstime();
-                assert(__quicklistCompressNode(node));
-                assert(__quicklistDecompressNode(node));
-                printf("Compress and decompress: %zu MB in %.2f seconds.\n",
-                       node->sz/1024/1024, (float)(mstime() - start) / 1000);
-            }
-
-            zfree(s);
-            zfree(node->prev);
-            zfree(node->next);
-            zfree(node->entry);
-            zfree(node);
-        }
-
-#if ULONG_MAX >= 0xffffffffffffffff
-        TEST("compress and decomress quicklist plain node large than UINT32_MAX") {
-            size_t sz = (1ull << 32);
-            unsigned char *s = zmalloc(sz);
-            randstring(s, sz);
-            memcpy(s, "helloworld", 10);
-            memcpy(s + sz - 10, "1234567890", 10);
-
-            quicklistNode *node = __quicklistCreateNode(QUICKLIST_NODE_CONTAINER_PLAIN, s, sz);
-
-            /* Just to avoid triggering the assertion in __quicklistCompressNode(),
-             * it disables the passing of quicklist head or tail node. */
-            node->prev = quicklistCreateNode();
-            node->next = quicklistCreateNode();
-
-            long long start = mstime();
-            assert(__quicklistCompressNode(node));
-            assert(__quicklistDecompressNode(node));
-            printf("Compress and decompress: %zu MB in %.2f seconds.\n",
-                   node->sz/1024/1024, (float)(mstime() - start) / 1000);
-
-            assert(memcmp(node->entry, "helloworld", 10) == 0);
-            assert(memcmp(node->entry + sz - 10, "1234567890", 10) == 0);
-            zfree(node->prev);
-            zfree(node->next);
-            zfree(node->entry);
-            zfree(node);
-        }
-#endif
-    }
-
-    if (!err)
-        printf("ALL TESTS PASSED!\n");
-    else
-        ERR("Sorry, not all tests passed!  In fact, %d tests failed.", err);
-
-    return err;
-}
-#endif
