+++ src/unit/test_kvstore.c
@@ -1,0 +1,144 @@
+#include <stdio.h>
+#include "testhelp.h"
+
+#define TEST(name) printf("test — %s\n", name);
+
+uint64_t hashTestCallback(const void *key) {
+    return dictGenHashFunction((unsigned char*)key, strlen((char*)key));
+}
+
+void freeTestCallback(dict *d, void *val) {
+    UNUSED(d);
+    zfree(val);
+}
+
+dictType KvstoreDictTestType = {
+    hashTestCallback,
+    NULL,
+    NULL,
+    NULL,
+    freeTestCallback,
+    NULL,
+    NULL
+};
+
+char *stringFromInt(int value) {
+    char buf[32];
+    int len;
+    char *s;
+
+    len = snprintf(buf, sizeof(buf), "%d",value);
+    s = zmalloc(len+1);
+    memcpy(s, buf, len);
+    s[len] = '\0';
+    return s;
+}
+
+/* ./valkey-server test kvstore */
+int kvstoreTest(int argc, char **argv, int flags) {
+    UNUSED(argc);
+    UNUSED(argv);
+    UNUSED(flags);
+
+    int i;
+    void *key;
+    dictEntry *de;
+    kvstoreIterator *kvs_it;
+    kvstoreDictIterator *kvs_di;
+
+    int didx = 0;
+    int curr_slot = 0;
+    kvstore *kvs1 = kvstoreCreate(&KvstoreDictTestType, 0, KVSTORE_ALLOCATE_DICTS_ON_DEMAND);
+    kvstore *kvs2 = kvstoreCreate(&KvstoreDictTestType, 0, KVSTORE_ALLOCATE_DICTS_ON_DEMAND | KVSTORE_FREE_EMPTY_DICTS);
+
+    TEST("Add 16 keys") {
+        for (i = 0; i < 16; i++) {
+            de = kvstoreDictAddRaw(kvs1, didx, stringFromInt(i), NULL);
+            assert(de != NULL);
+            de = kvstoreDictAddRaw(kvs2, didx, stringFromInt(i), NULL);
+            assert(de != NULL);
+        }
+        assert(kvstoreDictSize(kvs1, didx) == 16);
+        assert(kvstoreSize(kvs1) == 16);
+        assert(kvstoreDictSize(kvs2, didx) == 16);
+        assert(kvstoreSize(kvs2) == 16);
+    }
+
+    TEST("kvstoreIterator case 1: removing all keys does not delete the empty dict") {
+        kvs_it = kvstoreIteratorInit(kvs1);
+        while((de = kvstoreIteratorNext(kvs_it)) != NULL) {
+            curr_slot = kvstoreIteratorGetCurrentDictIndex(kvs_it);
+            key = dictGetKey(de);
+            assert(kvstoreDictDelete(kvs1, curr_slot, key) == DICT_OK);
+        }
+        kvstoreIteratorRelease(kvs_it);
+
+        dict *d = kvstoreGetDict(kvs1, didx);
+        assert(d != NULL);
+        assert(kvstoreDictSize(kvs1, didx) == 0);
+        assert(kvstoreSize(kvs1) == 0);
+    }
+
+    TEST("kvstoreIterator case 2: removing all keys will delete the empty dict") {
+        kvs_it = kvstoreIteratorInit(kvs2);
+        while((de = kvstoreIteratorNext(kvs_it)) != NULL) {
+            curr_slot = kvstoreIteratorGetCurrentDictIndex(kvs_it);
+            key = dictGetKey(de);
+            assert(kvstoreDictDelete(kvs2, curr_slot, key) == DICT_OK);
+        }
+        kvstoreIteratorRelease(kvs_it);
+
+        /* Make sure the dict was removed from the rehashing list. */
+        while (kvstoreIncrementallyRehash(kvs2, 1000)) {}
+
+        dict *d = kvstoreGetDict(kvs2, didx);
+        assert(d == NULL);
+        assert(kvstoreDictSize(kvs2, didx) == 0);
+        assert(kvstoreSize(kvs2) == 0);
+    }
+
+    TEST("Add 16 keys again") {
+        for (i = 0; i < 16; i++) {
+            de = kvstoreDictAddRaw(kvs1, didx, stringFromInt(i), NULL);
+            assert(de != NULL);
+            de = kvstoreDictAddRaw(kvs2, didx, stringFromInt(i), NULL);
+            assert(de != NULL);
+        }
+        assert(kvstoreDictSize(kvs1, didx) == 16);
+        assert(kvstoreSize(kvs1) == 16);
+        assert(kvstoreDictSize(kvs2, didx) == 16);
+        assert(kvstoreSize(kvs2) == 16);
+    }
+
+    TEST("kvstoreDictIterator case 1: removing all keys does not delete the empty dict") {
+        kvs_di = kvstoreGetDictSafeIterator(kvs1, didx);
+        while((de = kvstoreDictIteratorNext(kvs_di)) != NULL) {
+            key = dictGetKey(de);
+            assert(kvstoreDictDelete(kvs1, didx, key) == DICT_OK);
+        }
+        kvstoreReleaseDictIterator(kvs_di);
+
+        dict *d = kvstoreGetDict(kvs1, didx);
+        assert(d != NULL);
+        assert(kvstoreDictSize(kvs1, didx) == 0);
+        assert(kvstoreSize(kvs1) == 0);
+    }
+
+    TEST("kvstoreDictIterator case 2: removing all keys will delete the empty dict") {
+        kvs_di = kvstoreGetDictSafeIterator(kvs2, didx);
+        while((de = kvstoreDictIteratorNext(kvs_di)) != NULL) {
+            key = dictGetKey(de);
+            assert(kvstoreDictDelete(kvs2, didx, key) == DICT_OK);
+        }
+        kvstoreReleaseDictIterator(kvs_di);
+
+        dict *d = kvstoreGetDict(kvs2, didx);
+        assert(d == NULL);
+        assert(kvstoreDictSize(kvs2, didx) == 0);
+        assert(kvstoreSize(kvs2) == 0);
+    }
+
+    kvstoreRelease(kvs1);
+    kvstoreRelease(kvs2);
+    return 0;
+}
